test: Verify that ports won't collide
Add check to verify that we don't have too many tests, which with the current implementation of p2p_port() would mean the first tests in the list would risk getting the same ports as the last tests. This could be problematic the longer the first tests take, and the more jobs are run in parallel.

We currently have a limit of 415 tests, and the number of tests is about 100 less.

Based on investigation: https://github.com/bitcoin/bitcoin/issues/30390#issuecomment-2312554026