script: Remove shared_ptr-usage to clarify ownership
Greatly clarifies ownership of:
- miniscript::Node -> subs
- MiniscriptDescriptor -> m_node (now const).

Makes a lot of members in miniscript.h non-const in order to allow default move assignment operator(s) when Node is used directly in std::vector.

Also disables copying of Node as moving or Clone()-ing is meant to be used instead to avoid recursion and stack-overflow.