headerssync: No longer initiate additional headers syncs based of unrecognized block announcements
We now only attempt to initiate headers sync for a single peer in `PeerManagerImpl::SendMessages()`.

Next commit changes so that we start headers syncs from additional peers based off time instead.

Still flag the peer in the hope of giving it a higher chance of being selected for headers sync.

Reflection: A peer proposing a random block-hash not known to our currently considered best chain is not a sufficient heuristic for preferring it over other peers in a hostile environment. We need to consider peers whose behavior is similar to https://b10c.me/observations/06-linkinglion/ which are just listening.